update test_alexnet_save_and_load
Former-commit-id: 50e2628d5e6f1cf611cb0d37a794292a91817ade

+++ oneflow/python/test/serving/imagenet_record_dataset.py
@@ -102,7 +102,7 @@ class ImageNetRecordDataset(object):
             label_list.append(label)
 
         image_tensor = np.stack(image_list, axis=0)
-        label_tensor = np.array(label_list)
+        label_tensor = np.array(label_list, dtype=np.int32)
         self.num_read_batchs_ += 1
         return image_tensor, label_tensor
